refactor(web): extract apple JWKS verify and email sign-in eligibility helpers
